Added video entry addition functionality
Video entries can now be added on the main table. Checks are in place to
ensure a duplicate entry is not created.

Check added to ensure updating an existing entry can not be changed to
create a duplicate entry (i.e. two entries having the same course number
and filepath).